requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
joblib==1.3.2
numpy==1.26.0
scikit-learn==1.3.2
xgboost==1.7.6
python-multipart==0.0.6
PyPDF2==3.0.1
pytesseract==0.3.10
spacy==3.7.2
pillow==10.0.1
pandas==1.5.3
